==> 01_Data_Preprocessing.ipynb <==
# **Customer Lifetime Value (CLV) & Churn Prediction**

## **Problem Statement** : 
In the highly competitive non-contractual retail sector, customer acquisition costs are rising while customer loyalty is declining. This project addresses the challenge of ***Silent Attrition***—where customers simply stop buying without a formal cancellation. By leveraging transactional data, this project builds a predictive engine to identify high-value customers at risk of churning, enabling the business to proactively allocate marketing resources where they yield the highest ROI.

### **Objectives**
* Predict which customers are likely to churn in the next 6–12 months
* Estimate how much revenue each customer will generate in the future (CLV)
* Provide insights to retain high-value, high-risk customers
* Allow business managers to test strategies using a What-If simulation dashboard

### **Phases**
I will be executing this project in 7 phases in a structured way:
* **Phase 1 :** Data Cleaning & Transactional 
* **Phase 2 :** EDA - Exploratory Data Analysis
* **Phase 3 :** Feature Engineering - The RFM Transformation.
* **Phase 4 :** Probabilistic CLV Modeling - Predict future spending without using standard regression.
* **Phase 5 :** Churn Prediction
* **Phase 6 :** Dashboard with Discount vs Churn Impact
* **Phase 7:** Business Insights & Strategy

***Key Business Metrics***

* **CLV**                - Future revenue expected from a customer
* **Churn Probability**  - Likelihood that a customer will stop buying
* **Revenue at Risk**    - CLV × Churn probability

# First we need to import all the necessary libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import datetime as dt

from lifetimes.utils import summary_data_from_transaction_data, calibration_and_holdout_data
from lifetimes import BetaGeoFitter, GammaGammaFitter
from lifetimes.plotting import plot_period_transactions

import xgboost as xgb
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, confusion_matrix, mean_absolute_error, mean_squared_error

# Loading the dataset
df = pd.read_csv('online_retail_II.csv')
df.head()

# Checking shape of the dataset
df.shape

Our dataset has more than 10 lakhs of rows and 8 columns

# Explore Data Structure
df.info()

### ***Phase 1 :*** Data Cleaning & Transactional Logic
* Actions:
    * Standardizing columns name
    * Correcting Datatype
    * Removing rows with missing values.
    * Removing Duplicate orders.
    * Handling Returned/Cancelled orders.
    * Creating a '**total price**' (Quantity*Price) column.

First thing to do is make sure each column name is clean and standardize for avoiding any future complications.

df.columns = df.columns.str.lower().str.replace(" ", "_")
df.columns


The datatype of 'InvoiceDate' column is object. That's needed to be change into datettime datatype.

# Converting 'InvoiceDate' to datetime format
df['invoice_date'] = pd.to_datetime(df['invoicedate'])
df.drop(columns=['invoicedate'],errors = 'ignore', inplace=True)

df.info()

# Checking for missing values
df.isnull().sum()

# Calculating percentage of missing values
missing = df.isnull().mean().round(3)*100
print("Missing values (%):\n", missing)

Customer ID has 22.8% of missing values (around 2.5 lacs). We need to drop these rows since customer ID is crucial for CLV modelling.

df.dropna(subset = ['customer_id'], inplace=True)

# Checking for missing values
df.isnull().sum()

Now, there no missing values

# Checking for duplicate entries
df.duplicated().sum()

There are 26,749 duplicate rows which need to drop.

# Removing duplicate rows
df = df.drop_duplicates()

# Checking for shape after removing missing values
df.shape

Now I am left with 7,49,885 rows which are enough for model building.

# Let's also check for the date range of the dataset
print(df['invoice_date'].min(),' to ', df['invoice_date'].max())

Our dataset ranges from 1 Dec 2009 to 09 Dec 2011.

I also need to create a column as '*Total Price*' which shows the total amount spent by the customer on that order.

df['total_price'] = df['quantity'] * df['price']
df.head()

df.describe()

Seems like there are few negative values in the 'Quantity' column. Could be the returns order. Since these returned orders can affect the CLV model, I am creating a new dataframe to store these returned orders and removing these orders from the origional dataset.

returns_df = df[df['quantity'] < 0] # separate returns 
df = df[df['quantity'] > 0] # keep only purchases

# Final look at the cleaned dataset
df.head()

Now, the Dataset is cleaned from all the null and duplicate values, returned orders and correct datatypes. With this, I can move forward with the phase 2.

# Saving the cleaned dataset
df.to_csv("cleaned_transactions.csv", index=False)
print("File Saved!")


# def preprocess_data(df):
#     df = df.copy()
#     df.columns = df.columns.str.lower().str.replace(" ", "_")
#     df = df.drop_duplicates()
#     df = df[df['customer_id'].notnull()]
#     df['invoice_date'] = pd.to_datetime(df['invoice_date'], errors='coerce')
#     df = df[df['invoice_date'].notnull()]
#     df['total_price'] = df['quantity'] * df['price']
#     df['is_return'] = df['quantity'] < 0
#     df = df[df['is_return'] == False]
#     return df
# cleaned_df = preprocess_data(df)
