==> 03_Feature_Engineering.ipynb <==
# **Customer Lifetime Value (CLV) & Churn Prediction**

## ***Feature Engineering***
The goal of this phase is to transform cleaned transactional data into **customer-level, ML-ready features** that capture purchasing behavior, engagement, and value.

These engineered features will:
- Serve as inputs for **CLV estimation**
- Improve **churn prediction accuracy**
- Enable **business segmentation** (high-value vs low-value customers)

# Importing Necessary Libraries
import numpy as np
import pandas as pd

* Loading Dataset

# Importing dataset
df = pd.read_csv('cleaned_transactions.csv', parse_dates=['invoice_date'], dtype={'customer_id': str, 'invoice_id': str})
df.head()

### **Part A: RFM Calculation**
Creating RFM features to better understanding the customers' purchase behaviour as:
* Recency : number of days since a customer's last purchase   
* Frequency : number of unique invoices per customer(number of purchases by each customer)
* Monetary : total amount spent by a customer

# Setting reference date for recency calculation
ref_date = df['invoice_date'].max() + pd.Timedelta(days=1)

# Calculating RFM metrics
rfm = df.groupby('customer_id').agg({
    'invoice_date': lambda x: (ref_date - x.max()).days, # Recency
    'invoice': 'nunique',                                # Frequency
    'total_price':'sum'                                  # Monetary
}).reset_index()

# Renaming columns
rfm.columns = ['customer_id', 'recency','frequency', 'monetary']

# Displaying RFM table
rfm.head()

RFM features are calculated at a customer level by aggregating transactional data.

### **PART B — Additional Features**
I will be creating some additional features to RFM for better results.

* Average Order Value

rfm['avg_order_value'] = rfm['monetary']

* Purchase Interval

purchase_gaps = df.sort_values(['customer_id', 'invoice_date']).groupby('customer_id')['invoice_date'].diff().dt.days
interval_stats = purchase_gaps.groupby(df['customer_id']).agg(['mean','std'])

interval_stats.columns = ['purchase_interval_mean', 'purchase_interval_std']

rfm = rfm.merge(interval_stats, on='customer_id', how='left')

rfm.head()

* High Value Customer Flag

threshold = rfm['monetary'].quantile(0.75)
rfm['high_value_customer'] = rfm['monetary'] > threshold

* One-Time Buyer Flag

rfm['one_time_buyer'] = rfm['frequency'] == 1

* Time Since First Purchase

first_purchase = df.groupby('customer_id')['invoice_date'].min()
rfm = rfm.merge(first_purchase.rename('first_purchase_date'), on='customer_id')

rfm['customer_age_days'] = (ref_date - rfm['first_purchase_date']).dt.days
rfm.drop(columns=['first_purchase_date'], inplace=True)

### **Part C- Handling Missing/Skew Values**
Checking if there any missing values in newly created features.

rfm.isnull().sum()

Purchase interval may be null for single purchase customers. Its better to fill those null with 0.

rfm[['purchase_interval_mean', 'purchase_interval_std']] = rfm[['purchase_interval_mean', 'purchase_interval_std']].fillna(0)

* Log-transform skewed numeric columns

rfm['log_monetary'] = np.log1p(rfm['monetary'])
rfm['log_avg_order_values'] = np.log1p(rfm['avg_order_value'])

### **Final Dataset Check**

rfm.info()

* New dataset contains 12 columns and 5881 rows.
* Seems like all the columns have correct datatypes and no missing values.

rfm.describe().T

rfm.head()

With this, feature Engineering is also done.

This feature set forms the foundation for both CLV estimation and churn prediction models.

### **Exporting Dataset**

rfm.to_csv("final_rfm_features.csv", index=False)
