==> 04_CLV_Modeling.ipynb <==
# **Customer Lifetime Value (CLV) & Churn Prediction**

## ***CLV Modeling***
**Goal:** Predict customer purchase behavior and estimate 12-month CLV using probabilistic models.

**Models Used**
- BG/NBD - Purchase frequency prediction
- Gamma-Gamma - Monetary value prediction

# Importing Necessary Libraries
import numpy as np
import pandas as pd

from lifetimes import BetaGeoFitter, GammaGammaFitter
from lifetimes.utils import summary_data_from_transaction_data

# Loading dataset
df = pd.read_csv('cleaned_transactions.csv', parse_dates=['invoice_date'], dtype={'customer_id': str, 'invoice_id': str})
df.head()

### **Section 1:** Prepare Data for CLV Models

Creating summary of data for BG/NBD Model

summary = summary_data_from_transaction_data(
    df,
    customer_id_col='customer_id',
    datetime_col='invoice_date',
    monetary_value_col='total_price',
    observation_period_end=df['invoice_date'].max()
    )

summary.head()

BG/NBD Model requires frequency > 0. There, need to remove the rows with frequency = 0.

summary = summary[summary['frequency'] > 0]

### **Section 2:** BG/NBD Model (Purchase Frequency)
BG/NBD models the probability that a customer is still active and estimates how often they will purchase in the future based on historical behavior.

bgf = BetaGeoFitter(penalizer_coef=0.01)
bgf.fit(
    summary['frequency'],
    summary['recency'],
    summary['T']
)

Predicting purchases in the next 12th months(=365 days)

summary['predicted_purchases_12m'] = bgf.conditional_expected_number_of_purchases_up_to_time(
    365,
    summary['frequency'],
    summary['recency'],
    summary['T']
)

### **Section 3:** Gamma-Gamma Model (Monetary Value)
Gamma-Gamma estimates the expected monetary value of a transaction assuming spending behavior is independent of purchase frequency.

ggf = GammaGammaFitter(penalizer_coef = 0.01)
ggf.fit(
    summary['frequency'],
    summary['monetary_value']
)

Predicting expected average order value.

summary['expected_avg_order_value'] = ggf.conditional_expected_average_profit(
    summary['frequency'],
    summary['monetary_value']
)

### **Section 4:** Calculate 12 Month CLV
CLV represents the discounted expected revenue a customer will generate over the next 12 months.

Now, combining both BD/NBG and Gamma Gamma Model, I will be calculating the CLV of each customer for the next 12 months(These models assume customer purchase behavior remains stable over the prediction horizon).

summary['clv_12m'] = ggf.customer_lifetime_value(
    bgf,
    summary['frequency'],
    summary['recency'],
    summary['T'],
    summary['monetary_value'],
    time=12,  # months
    freq='D',  # daily data
    discount_rate=0.01
)

### **Section 6:** Inspecting Results

summary[['predicted_purchases_12m',
         'expected_avg_order_value',
         'clv_12m']].describe().T

* Top 10 Customers by CLV

summary.sort_values(by='clv_12m', ascending=False).head(10)

These customers most contributes the most in the company, therefore they can be our VIP customers.

#### Saving the CLV Output
This output will be used in:
* Churn Model
* Power BI Dashboard
* What-if Simulations

# Exporting CLV Predictions
summary = summary.reset_index()
summary['customer_id'] = summary['customer_id'].astype(str)

summary.reset_index().to_csv("clv_predictions.csv", index=False)

summary.head()

summary.isnull().sum()
