==> 05_Churn_Modeling.ipynb <==
# **Customer Lifetime Value (CLV) & Churn Prediction**
## ***Churn Prediction Modeling***
**Goal:** Predict the probability that a customer will churn using behavioral features and CLV.

This model enables:
- Proactive retention strategies
- Prioritization of high-value, high-risk customers
- Integration with a What-If dashboard

# Importing necessary libraries
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score, classification_report, confusion_matrix
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.ensemble import RandomForestClassifier

First, I will import two datasets: "final_rfm_features.csv" and "clv_predictions.csv", and merge then into a single dataframe to predict customer churn. 

# Importing datasets
features = pd.read_csv("final_rfm_features.csv", dtype={'customer_id': str})
clv = pd.read_csv("clv_predictions.csv", dtype={'customer_id': str})

# Merging CLV predictions with RFM features
df = features.merge(
    clv[['customer_id', 'clv_12m']],
    on='customer_id',
    how='left'
)

df.head()

Customer identifiers were standardized across datasets to ensure correct joins between feature, CLV, and churn tables.

### **Section 1:** Defining Churn Label

Churn is defined as customer inactivity beyond 90 days, a commonly used proxy in non-contractual retail businesses.

# Setting churn threshold to 90 for customers to hadn't bought anything in the last 90 days
churn_threshold = 90

df['churn'] = (df['recency'] > churn_threshold).astype(int)
df['churn'].value_counts(normalize=True)

### **Section 2:** Selecting Features for ML model

# Removing identifier column
drop_cols = ['customer_id']

# Defining features and target variable
X = df.drop(columns = drop_cols + ['churn'])
y = df['churn']

# Checking feature set
X.head()

#### **Section 3:** Train-Test Split

# Splitting data into training and testing sets with stratification
X_train, X_test, y_train, y_test = train_test_split(
    X,y, test_size=0.25, random_state=42, stratify=y
)

#### **Section 4:** Building ML Pipeline

pipeline = Pipeline([
    ('scaler', StandardScaler()),
    ('model', RandomForestClassifier(
        n_estimators=300,
        max_depth=8,
        class_weight='balanced',
        random_state=42
    ))
])

* Random Forest was chosen for its ability to capture non-linear relationships and handle mixed feature types with minimal preprocessing.

# Fitting the pipeline to the training data
pipeline.fit(X_train, y_train)

#### **Section 5:** Evaluating the Model

# Making predictions and evaluating the model
y_pred = pipeline.predict(X_test)
y_proba = pipeline.predict_proba(X_test)[:, 1]

# Evaluating model performance
print("ROC AUC Score:", roc_auc_score(y_test, y_proba))
print("\nClassification Report:\n", classification_report(y_test, y_pred))

Also, checking the confusion matrix.

# Displaying the confusion matrix
confusion_matrix(y_test, y_pred)

The model is showing ROC-AUC score = 1, with perfect precision, recall and accuracy; which indicates data leakage.
Let's check what happen:
* I set recency > 90 as churn indicator
* But we are also using recency based features for churn prediction, causing data leakage.
* That's why model is predicting: if recency > 90, Churn; else Not churn.

Although recency is a strong indicator of churn, it needs to be excluded from the churn
prediction feature set because churn was defined using recency.
Including recency would result in target leakage and unrealistically high model performance.

#### **Section 6:** Fixing Data Leakage
* Step 1: Removing recency and recency deried features from features set

leakage_cols = ['recency', 'purchase_interval_mean', 'purchase_interval_std']
X = X.drop(columns=leakage_cols)

# Checking feature set again
X.head()

These features looks safe to train the model.

* Step 2:  Splitting data into train and test set.

# Splitting data into training and testing sets again
X_train, X_test, y_train, y_test = train_test_split(
    X,y, test_size=0.25, random_state=42, stratify=y
)

* Step 3: Re-Training the model

pipeline.fit(X_train, y_train)
y_proba = pipeline.predict_proba(X_test)[:,1]

print("ROC AUC Score after removing leakage features:", roc_auc_score(y_test, y_proba))

#### Final Churn Model Performance

After correcting for data leakage, the churn model achieved:

- **ROC-AUC: 0.948**
- Strong recall for high-risk customers
- Stable probability estimates suitable for ranking and prioritization

This confirms the model’s effectiveness in identifying churn prone customers without relying on future information.

#### **Section 7:** Feature Importance
It is imporatant to check which factors are causing customers to churn, so that business can make strategies to retain customers.

# Analyzing feature importance
importances = pipeline.named_steps['model'].feature_importances_

# Creating a DataFrame for feature importance
feature_importance = (
    pd.DataFrame({
        'features' : X.columns,
        'importance' : importances
    }).sort_values('importance', ascending=False)
)

# Displaying top 10 important features
feature_importance.head(10)

#### **Section 8:** Predicting Churn Probability for all customers

# Adding churn probability to the original dataframe
df['churn_probability'] = pipeline.predict_proba(X)[:,1]

Also, putting each customers into different segments based on their risk level.

# Segmenting customers based on churn risk
df['risk_segment'] = pd.cut(
    df['churn_probability'],
    bins=[0,0.3,0.6,1.0],
    labels=['low risk', 'medium risk', 'high risk']
)

There seems to be some missing values in "clv_12m" column. Let's check them.

# Checking for missing values
df.isnull().sum()

There are 1692 missing values in clv_12m column. That can be because in previous phase, we only used data of customers which have **frequency > 0** (since clv can't be calculated for frequency =0). That why we don't have clv for customers with frequency = 0. These missing values can be filled with 0.

# Filling missing values with 0
df["clv_12m"] = df["clv_12m"].fillna(0)

#### **Section 9:** Revenue at Risk

# Calculating how much revenue is at risk due to potential churn
df['revenue_at_risk'] = df['churn_probability'] * df['clv_12m']

# Displaying sample of final dataframe with churn probabilities and revenue at risk
df[['customer_id', 'churn_probability','clv_12m','revenue_at_risk']].head(10)

Customers with high churn probability and high CLV represent the highest revenue risk and should be prioritized for retention campaigns.

* Exporting Dataset

# Saving the final dataframe with churn predictions
df.to_csv("churn_predictions.csv", index=False)

### Modeling Notes & Limitations

- Churn was defined using an inactivity-based rule due to the non-contractual nature of the business.
- Recency-based features were intentionally excluded from the churn model to prevent data leakage.
- The resulting churn probabilities represent relative risk scores rather than absolute churn guarantees.

These outputs are suitable for prioritization and decision-making in retention strategies.
